stopped resizing resume image, losing too much quality

+++ deploy.sh
@@ -6,24 +6,25 @@ cd "$(dirname "$0")"
 # Font subset
 ./get-all-chars.py space.ttf
 
+# Get Resume JPG, convert to webp, and resize
+wget https://raw.githubusercontent.com/cohenchris/resume/master/ChrisCohen_resume.jpg
+cwebp -q 25 ChrisCohen_resume.jpg -o static/images/ChrisCohen_resume.webp
+rm ChrisCohen_resume.jpg
+
 # Sync music listening progress metadata
 python3 ./website-listening-progress-json.py
 
-# Get Resume JPG, convert to webp, and resize
-wget https://raw.githubusercontent.com/cohenchris/resume/master/ChrisCohen_resume.jpg
-cwebp -q 80 ChrisCohen_resume.jpg -o static/images/ChrisCohen_resume.webp
-cwebp -resize 800 1131 static/images/ChrisCohen_resume.webp -o static/images/ChrisCohen_resume.webp
-rm ChrisCohen_resume.jpg
+if [ "$1" != "test" ]; then
+  # Generate and deploy to server
+  hugo
 
-# Generate and deploy to server
-hugo
+  mv public/ html/
 
-mv public/ html/
+  rm -r /home/phrog/server/config/swag/www/chriscohen.dev/html
 
-rm -r /home/phrog/server/config/swag/www/chriscohen.dev/html
+  mv html /home/phrog/server/config/swag/www/chriscohen.dev/
 
-mv html /home/phrog/server/config/swag/www/chriscohen.dev/
-
-cd $starting_dir
+  cd $starting_dir
+fi
 
 echo "Don't forget to commit!"
